Fix dll copy error

+++ Database/FormRegister.cs
@@ -26,7 +26,10 @@ namespace Database
             {
                 path = path.Remove(path.LastIndexOf('\\'));
             }
-            File.Copy(path + @"\dll\SQLite.Interop.dll", Directory.GetCurrentDirectory()+ @"\SQLite.Interop.dll");
+            if (!(File.Exists(Directory.GetCurrentDirectory() + @"\SQLite.Interop.dll")))
+            {
+                File.Copy(path + @"\dll\SQLite.Interop.dll", Directory.GetCurrentDirectory() + @"\SQLite.Interop.dll");
+            }
         }
 
         private void btnConnect_Click(object sender, EventArgs e)
@@ -128,56 +131,55 @@ namespace Database
         private void btnCreate_Click(object sender, EventArgs e)
         {
             #region SQLServer
-            string dataBase = Application.StartupPath + "DBSQLServer.sdf";
-            string strConnection = @"DataSource = " + dataBase + "; Password = 'admin'";
+            //string dataBase = Application.StartupPath + "DBSQLServer.sdf";
+            //string strConnection = @"DataSource = " + dataBase + "; Password = 'admin'";
 
-            SqlCeConnection connection = new SqlCeConnection(strConnection);
-            try
-            {
-                connection.Open();
-                SqlCeCommand command = new SqlCeCommand();
-                command.Connection = connection;
-                command.CommandText = "CREATE TABLE people (id INT NOT NULL PRIMARY KEY, name NVARCHAR(50), email NVARCHAR(50))";
-                command.ExecuteNonQuery();
-
-                labelResult.Text = "Table CREATED successfully! [SQL Server CE]";
-                command.Dispose();
-
-            }
-            catch (Exception ex)
-            {
-                labelResult.Text = "Error!\n" + ex;
-
-            }
-            finally { connection.Close(); }
-
-
-            #endregion
-
-            #region SQLite
-
-            //string dataBase = Application.StartupPath + "DBSQLite.db";
-            //string strConnection = @"Data Source = " + dataBase + "; Version = 3";
-
-            //SQLiteConnection connection = new SQLiteConnection(strConnection);
+            //SqlCeConnection connection = new SqlCeConnection(strConnection);
             //try
             //{
             //    connection.Open();
-            //    SQLiteCommand command = new SQLiteCommand();
+            //    SqlCeCommand command = new SqlCeCommand();
             //    command.Connection = connection;
             //    command.CommandText = "CREATE TABLE people (id INT NOT NULL PRIMARY KEY, name NVARCHAR(50), email NVARCHAR(50))";
             //    command.ExecuteNonQuery();
 
-            //    labelResult.Text = "Table CREATED successfully! [SQLite]";
+            //    labelResult.Text = "Table CREATED successfully! [SQL Server CE]";
             //    command.Dispose();
 
             //}
             //catch (Exception ex)
             //{
-            //    labelResult.Text = "Error! " + ex;
+            //    labelResult.Text = "Error!\n" + ex;
 
             //}
             //finally { connection.Close(); }
+            #endregion
+
+            #region SQLite
+
+            string dataBase = Application.StartupPath + "DBSQLite.db";
+            string strConnection = @"Data Source = " + dataBase + "; Version = 3";
+
+            SQLiteConnection connection = new SQLiteConnection(strConnection);
+
+            try
+            {
+                connection.Open();
+                SQLiteCommand command = new SQLiteCommand();
+                command.Connection = connection;
+                command.CommandText = "CREATE TABLE people (id INT NOT NULL PRIMARY KEY, name NVARCHAR(50), email NVARCHAR(50))";
+                command.ExecuteNonQuery();
+
+                labelResult.Text = "Table CREATED successfully! [SQLite]";
+                command.Dispose();
+
+            }
+            catch (Exception ex)
+            {
+                labelResult.Text = "Error! " + ex;
+
+            }
+            finally { connection.Close(); }
 
             #endregion
 
@@ -212,47 +214,14 @@ namespace Database
         {
             #region SQLServer
 
-            string dataBase = Application.StartupPath + "DBSQLServer.sdf"; //Local
-            string strConnection = @"DataSource = " + dataBase + "; Password = 'admin'"; //Parametros
+            //string dataBase = Application.StartupPath + "DBSQLServer.sdf"; //Local
+            //string strConnection = @"DataSource = " + dataBase + "; Password = 'admin'"; //Parametros
 
-            SqlCeConnection connection = new SqlCeConnection(strConnection);
-            try
-            {
-                connection.Open();
-                SqlCeCommand command = new SqlCeCommand();
-                command.Connection = connection;
-
-                int id = new Random(DateTime.Now.Millisecond).Next(0, 1000);
-                string name = txtName.Text;
-                string email = txtEmail.Text;
-
-                command.CommandText = $"INSERT INTO people VALUES ({id}, '{name}', '{email}')";
-                command.ExecuteNonQuery();
-
-                labelResult.Text = "Data INSERTED successfully! [SQL Server CE]";
-                command.Dispose();
-
-            }
-            catch (Exception ex)
-            {
-                labelResult.Text = ex.ToString();
-
-            }
-            finally { connection.Close(); txtName.Text = ""; txtEmail.Text = ""; }
-
-
-            #endregion
-
-            #region SQLite
-
-            //string dataBase = Application.StartupPath + "DBSQLite.db";
-            //string strConnection = @"Data Source = " + dataBase + "; Version = 3";
-
-            //SQLiteConnection connection = new SQLiteConnection(strConnection);
+            //SqlCeConnection connection = new SqlCeConnection(strConnection);
             //try
             //{
             //    connection.Open();
-            //    SQLiteCommand command = new SQLiteCommand();
+            //    SqlCeCommand command = new SqlCeCommand();
             //    command.Connection = connection;
 
             //    int id = new Random(DateTime.Now.Millisecond).Next(0, 1000);
@@ -262,7 +231,7 @@ namespace Database
             //    command.CommandText = $"INSERT INTO people VALUES ({id}, '{name}', '{email}')";
             //    command.ExecuteNonQuery();
 
-            //    labelResult.Text = "Data INSERTED successfully! [SQLite]";
+            //    labelResult.Text = "Data INSERTED successfully! [SQL Server CE]";
             //    command.Dispose();
 
             //}
@@ -272,6 +241,39 @@ namespace Database
 
             //}
             //finally { connection.Close(); txtName.Text = ""; txtEmail.Text = ""; }
+
+
+            #endregion
+
+            #region SQLite
+
+            string dataBase = Application.StartupPath + "DBSQLite.db";
+            string strConnection = @"Data Source = " + dataBase + "; Version = 3";
+
+            SQLiteConnection connection = new SQLiteConnection(strConnection);
+            try
+            {
+                connection.Open();
+                SQLiteCommand command = new SQLiteCommand();
+                command.Connection = connection;
+
+                int id = new Random(DateTime.Now.Millisecond).Next(0, 1000);
+                string name = txtName.Text;
+                string email = txtEmail.Text;
+
+                command.CommandText = $"INSERT INTO people VALUES ({id}, '{name}', '{email}')";
+                command.ExecuteNonQuery();
+
+                labelResult.Text = "Data INSERTED successfully! [SQLite]";
+                command.Dispose();
+
+            }
+            catch (Exception ex)
+            {
+                labelResult.Text = ex.ToString();
+
+            }
+            finally { connection.Close(); txtName.Text = ""; txtEmail.Text = ""; }
 
 
             #endregion
@@ -312,53 +314,13 @@ namespace Database
         {
             #region SQLServer
 
-            list.Rows.Clear();
-            labelResult.Text = "";
-
-            string dataBase = Application.StartupPath + "DBSQLServer.sdf";
-            string strConnection = @"DataSource = " + dataBase + "; Password = 'admin'";
-
-            SqlCeConnection connection = new SqlCeConnection(strConnection);
-            try
-            {
-                string query = "SELECT * FROM people";
-
-                if (txtName.Text != "")
-                {
-                    query = "SELECT * FROM people WHERE name LIKE '" + txtName.Text + "'";
-                }
-
-                DataTable data = new DataTable();
-                SqlCeDataAdapter adapter = new SqlCeDataAdapter(query, strConnection);
-
-                connection.Open();
-
-                adapter.Fill(data);
-                foreach (DataRow line in data.Rows)
-                {
-                    list.Rows.Add(line.ItemArray);
-                }
-            }
-            catch (Exception ex)
-            {
-                list.Rows.Clear();
-                labelResult.Text = ex.ToString();
-
-            }
-            finally { connection.Close(); txtName.Text = ""; txtEmail.Text = ""; }
-
-
-            #endregion
-
-            #region SQLite
-
             //list.Rows.Clear();
             //labelResult.Text = "";
 
-            //string dataBase = Application.StartupPath + "DBSQLite.db";
-            //string strConnection = @"Data Source = " + dataBase + "; Version = 3";
+            //string dataBase = Application.StartupPath + "DBSQLServer.sdf";
+            //string strConnection = @"DataSource = " + dataBase + "; Password = 'admin'";
 
-            //SQLiteConnection connection = new SQLiteConnection(strConnection);
+            //SqlCeConnection connection = new SqlCeConnection(strConnection);
             //try
             //{
             //    string query = "SELECT * FROM people";
@@ -369,7 +331,7 @@ namespace Database
             //    }
 
             //    DataTable data = new DataTable();
-            //    SQLiteDataAdapter adapter = new SQLiteDataAdapter(query, strConnection);
+            //    SqlCeDataAdapter adapter = new SqlCeDataAdapter(query, strConnection);
 
             //    connection.Open();
 
@@ -386,6 +348,46 @@ namespace Database
 
             //}
             //finally { connection.Close(); txtName.Text = ""; txtEmail.Text = ""; }
+
+
+            #endregion
+
+            #region SQLite
+
+            list.Rows.Clear();
+            labelResult.Text = "";
+
+            string dataBase = Application.StartupPath + "DBSQLite.db";
+            string strConnection = @"Data Source = " + dataBase + "; Version = 3";
+
+            SQLiteConnection connection = new SQLiteConnection(strConnection);
+            try
+            {
+                string query = "SELECT * FROM people";
+
+                if (txtName.Text != "")
+                {
+                    query = "SELECT * FROM people WHERE name LIKE '" + txtName.Text + "'";
+                }
+
+                DataTable data = new DataTable();
+                SQLiteDataAdapter adapter = new SQLiteDataAdapter(query, strConnection);
+
+                connection.Open();
+
+                adapter.Fill(data);
+                foreach (DataRow line in data.Rows)
+                {
+                    list.Rows.Add(line.ItemArray);
+                }
+            }
+            catch (Exception ex)
+            {
+                list.Rows.Clear();
+                labelResult.Text = ex.ToString();
+
+            }
+            finally { connection.Close(); txtName.Text = ""; txtEmail.Text = ""; }
 
             #endregion
 
@@ -433,45 +435,14 @@ namespace Database
         {
             #region SQLServer
 
-            string dataBase = Application.StartupPath + "DBSQLServer.sdf"; //Local
-            string strConnection = @"DataSource = " + dataBase + "; Password = 'admin'"; //Parametros
+            //string dataBase = Application.StartupPath + "DBSQLServer.sdf"; //Local
+            //string strConnection = @"DataSource = " + dataBase + "; Password = 'admin'"; //Parametros
 
-            SqlCeConnection connection = new SqlCeConnection(strConnection);
-            try
-            {
-                connection.Open();
-                SqlCeCommand command = new SqlCeCommand();
-                command.Connection = connection;
-
-                int id = (int)list.SelectedRows[0].Cells[0].Value;
-
-                command.CommandText = $"DELETE FROM people WHERE id='{id}'";
-                command.ExecuteNonQuery();
-
-                labelResult.Text = "Data DELETED successfully! [SQL Server CE]";
-                command.Dispose();
-
-            }
-            catch (Exception ex)
-            {
-                labelResult.Text = ex.ToString();
-
-            }
-            finally { connection.Close(); txtName.Text = ""; txtEmail.Text = ""; }
-
-
-            #endregion
-
-            #region SQLite
-
-            //string dataBase = Application.StartupPath + "DBSQLite.db";
-            //string strConnection = @"Data Source = " + dataBase + "; Version = 3";
-
-            //SQLiteConnection connection = new SQLiteConnection(strConnection);
+            //SqlCeConnection connection = new SqlCeConnection(strConnection);
             //try
             //{
             //    connection.Open();
-            //    SQLiteCommand command = new SQLiteCommand();
+            //    SqlCeCommand command = new SqlCeCommand();
             //    command.Connection = connection;
 
             //    int id = (int)list.SelectedRows[0].Cells[0].Value;
@@ -479,7 +450,7 @@ namespace Database
             //    command.CommandText = $"DELETE FROM people WHERE id='{id}'";
             //    command.ExecuteNonQuery();
 
-            //    labelResult.Text = "Data DELETED successfully! [SQLite]";
+            //    labelResult.Text = "Data DELETED successfully! [SQL Server CE]";
             //    command.Dispose();
 
             //}
@@ -489,6 +460,37 @@ namespace Database
 
             //}
             //finally { connection.Close(); txtName.Text = ""; txtEmail.Text = ""; }
+
+
+            #endregion
+
+            #region SQLite
+
+            string dataBase = Application.StartupPath + "DBSQLite.db";
+            string strConnection = @"Data Source = " + dataBase + "; Version = 3";
+
+            SQLiteConnection connection = new SQLiteConnection(strConnection);
+            try
+            {
+                connection.Open();
+                SQLiteCommand command = new SQLiteCommand();
+                command.Connection = connection;
+
+                int id = (int)list.SelectedRows[0].Cells[0].Value;
+
+                command.CommandText = $"DELETE FROM people WHERE id='{id}'";
+                command.ExecuteNonQuery();
+
+                labelResult.Text = "Data DELETED successfully! [SQLite]";
+                command.Dispose();
+
+            }
+            catch (Exception ex)
+            {
+                labelResult.Text = ex.ToString();
+
+            }
+            finally { connection.Close(); txtName.Text = ""; txtEmail.Text = ""; }
 
 
             #endregion
@@ -527,46 +529,14 @@ namespace Database
         {
             #region SQLServer
 
-            string dataBase = Application.StartupPath + "DBSQLServer.sdf"; //Local
-            string strConnection = @"DataSource = " + dataBase + "; Password = 'admin'"; //Parametros
+            //string dataBase = Application.StartupPath + "DBSQLServer.sdf"; //Local
+            //string strConnection = @"DataSource = " + dataBase + "; Password = 'admin'"; //Parametros
 
-            SqlCeConnection connection = new SqlCeConnection(strConnection);
-            try
-            {
-                connection.Open();
-                SqlCeCommand command = new SqlCeCommand();
-                command.Connection = connection;
-
-                int id = (int)list.SelectedRows[0].Cells[0].Value;
-                string query = $"UPDATE people SET name='{txtName.Text}', email='{txtEmail.Text}' WHERE id LIKE '{id}'";
-
-                command.CommandText = query;
-                command.ExecuteNonQuery();
-
-                labelResult.Text = "Data UPDATED successfully! [SQL Server CE]";
-                command.Dispose();
-
-            }
-            catch (Exception ex)
-            {
-                labelResult.Text = ex.ToString();
-
-            }
-            finally { connection.Close(); txtName.Text = ""; txtEmail.Text = ""; }
-
-
-            #endregion
-
-            #region SQLite
-
-            //string dataBase = Application.StartupPath + "DBSQLite.db";
-            //string strConnection = @"Data Source = " + dataBase + "; Version = 3";
-
-            //SQLiteConnection connection = new SQLiteConnection(strConnection);
+            //SqlCeConnection connection = new SqlCeConnection(strConnection);
             //try
             //{
             //    connection.Open();
-            //    SQLiteCommand command = new SQLiteCommand();
+            //    SqlCeCommand command = new SqlCeCommand();
             //    command.Connection = connection;
 
             //    int id = (int)list.SelectedRows[0].Cells[0].Value;
@@ -575,7 +545,7 @@ namespace Database
             //    command.CommandText = query;
             //    command.ExecuteNonQuery();
 
-            //    labelResult.Text = "Data UPDATED successfully! [SQLite]";
+            //    labelResult.Text = "Data UPDATED successfully! [SQL Server CE]";
             //    command.Dispose();
 
             //}
@@ -585,6 +555,38 @@ namespace Database
 
             //}
             //finally { connection.Close(); txtName.Text = ""; txtEmail.Text = ""; }
+
+
+            #endregion
+
+            #region SQLite
+
+            string dataBase = Application.StartupPath + "DBSQLite.db";
+            string strConnection = @"Data Source = " + dataBase + "; Version = 3";
+
+            SQLiteConnection connection = new SQLiteConnection(strConnection);
+            try
+            {
+                connection.Open();
+                SQLiteCommand command = new SQLiteCommand();
+                command.Connection = connection;
+
+                int id = (int)list.SelectedRows[0].Cells[0].Value;
+                string query = $"UPDATE people SET name='{txtName.Text}', email='{txtEmail.Text}' WHERE id LIKE '{id}'";
+
+                command.CommandText = query;
+                command.ExecuteNonQuery();
+
+                labelResult.Text = "Data UPDATED successfully! [SQLite]";
+                command.Dispose();
+
+            }
+            catch (Exception ex)
+            {
+                labelResult.Text = ex.ToString();
+
+            }
+            finally { connection.Close(); txtName.Text = ""; txtEmail.Text = ""; }
 
 
             #endregion
